move to `message_extensions` class
