--- python/packages/ai/teams/message_preview_action.py
@@ -1,8 +1,0 @@
-"""
-Copyright (c) Microsoft Corporation. All rights reserved.
-Licensed under the MIT License.
-"""
-
-from typing import Literal
-
-MessagePreviewAction = Literal["edit", "send"]
